refactor: 테스트 공통 코드 분리
